fixed a bug on copy types, FOR REAL
the thing is, the type of [copy [println]] *isn't*
['X → ['a 'Y → 'Y] ['a 'Y → 'Y] 'X]!! that implies they have to be
called with the same arguments, or whatever. the type is actually
['X → ['a 'Y → 'Y] ['b 'Z → 'Z] 'X]. So I had to add a way to formalize
this notion of "types that used to be the same but aren't anymore."
Henceforth if you have two of the same type variable on the right hand
side of a function type you're building in C, you should probably wrap
them in diverge().