some ways to restrict the types of 'apply'
and a very nice build-your-own-function system! with meta-functions!
read tests/0test! it's cool! and there are comments!

man, once we get the stdlib going and implemented IN THE LANGUAGE ITSELF,
this is going to be SO COOL!
as it is you kind of do have to copy-paste function definitions
everywhere you want to use them though :/