change separator char to |
it doesn't really look right as a semicolon, since the statements are
still more deeply connected, and a pipe gives much more of the right
impression -- we're taking the output of the last little function and
piping it into the next one, like a UNIX pipe.

e.g. it makes much more sense to me to read:
        { 1 2 3 4 } | print len?
(we're taking this list and piping it to the function that
 non-destructively prints its length)
than:
        { 1 2 3 4 } ; print len?
(which looks too imperative -- we're evaluating this list, then...
 printing a length? what does that mean?)