start making it postfix
prefix by default is too confusing. better to provide syntax for things
that work better as prefix (eg if, while), and leave everything else
left-to-right. idk i've felt committed to the prefix thing for a while
but I can just feel the language fighting against it. now I get why
everything is postfix. HOWEVER the whole "condition [then] [else] if"
thing is awful, so I'm providing syntax to avoid stuff like that.