for syntax extensibility/consistency, actually require 'func'
OKAY I'm sorry to keep changing the syntax (said to nobody in particular)
but I think this is the best way to do this -- now we can always parse
the first word in the declaration as specifying the 'type of thing'
being declared, which leaves us a lot of leeway to add new keywords here
without conflicting with user-defined names. also maybe maybe we could
even allow custom declaration macro types? idk lets not get our hopes up