Standard library file added: lib/std.alma
Turns out I had to do a lot of rejiggering of the execution model to
make this work. Which then exposed a lot of bad design decisions (like
the func registry not working the way it was supposed to.) So hopefully
this makes a bit more sense! I think I got all the memory leaks.